+++ lib/groupdocs/api.rb
@@ -1,2 +1,3 @@
+require 'groupdocs/api/helpers'
 require 'groupdocs/api/entity'
 require 'groupdocs/api/request'