+++ spec/groupdocs/document/annotation/reviewer_spec.rb
@@ -45,13 +45,9 @@ describe GroupDocs::Document::Annotation::Reviewer do
     end
   end
 
-  it { should respond_to(:emailAddress)  }
-  it { should respond_to(:emailAddress=) }
-  it { should respond_to(:FullName)      }
-  it { should respond_to(:FullName=)     }
+  it { should have_accessor(:emailAddress) }
+  it { should have_accessor(:FullName)     }
 
-  it { should have_alias(:email_address, :emailAddress)   }
-  it { should have_alias(:email_address=, :emailAddress=) }
-  it { should have_alias(:full_name, :FullName)           }
-  it { should have_alias(:full_name=, :FullName=)         }
+  it { should alias_accessor(:email_address, :emailAddress) }
+  it { should alias_accessor(:full_name, :FullName)         }
 end