+++ spec/groupdocs/document/rectangle_spec.rb
@@ -4,17 +4,11 @@ describe GroupDocs::Document::Rectangle do
 
   it_behaves_like GroupDocs::Api::Entity
 
-  it { should respond_to(:x)       }
-  it { should respond_to(:x=)      }
-  it { should respond_to(:y)       }
-  it { should respond_to(:y=)      }
-  it { should respond_to(:width)   }
-  it { should respond_to(:width=)  }
-  it { should respond_to(:height)  }
-  it { should respond_to(:height=) }
+  it { should have_accessor(:x)      }
+  it { should have_accessor(:y)      }
+  it { should have_accessor(:width)  }
+  it { should have_accessor(:height) }
 
-  it { should have_alias(:w, :width)    }
-  it { should have_alias(:w=, :width=)  }
-  it { should have_alias(:h, :height)   }
-  it { should have_alias(:h=, :height=) }
+  it { should alias_accessor(:w, :width)  }
+  it { should alias_accessor(:h, :height) }
 end