+++ spec/groupdocs/signature/envelope/log_spec.rb
@@ -4,20 +4,12 @@ describe GroupDocs::Signature::Envelope::Log do
 
   it_behaves_like GroupDocs::Api::Entity
 
-  it { should respond_to(:id)             }
-  it { should respond_to(:id=)            }
-  it { should respond_to(:date)           }
-  it { should respond_to(:date=)          }
-  it { should respond_to(:userName)       }
-  it { should respond_to(:userName=)      }
-  it { should respond_to(:action)         }
-  it { should respond_to(:action=)        }
-  it { should respond_to(:remoteAddress)  }
-  it { should respond_to(:remoteAddress=) }
+  it { should have_accessor(:id)            }
+  it { should have_accessor(:date)          }
+  it { should have_accessor(:userName)      }
+  it { should have_accessor(:action)        }
+  it { should have_accessor(:remoteAddress) }
 
-  it { should have_alias(:user_name,       :userName)       }
-  it { should have_alias(:user_name=,      :userName=)      }
-  it { should have_alias(:remote_address,  :remoteAddress)  }
-  it { should have_alias(:remote_address=, :remoteAddress=) }
-
+  it { should alias_accessor(:user_name, :userName)            }
+  it { should alias_accessor(:remote_address,  :remoteAddress) }
 end