+++ spec/support/shared_examples/signature/shared/entity_fields.rb
@@ -1,52 +1,31 @@
 shared_examples_for GroupDocs::Signature::EntityFields do
 
-  it { should respond_to(:id)                }
-  it { should respond_to(:id=)               }
-  it { should respond_to(:name)              }
-  it { should respond_to(:name=)             }
-  it { should respond_to(:ownerId)           }
-  it { should respond_to(:ownerId=)          }
-  it { should respond_to(:ownerGuid)         }
-  it { should respond_to(:ownerGuid=)        }
-  it { should respond_to(:reminderTime)      }
-  it { should respond_to(:reminderTime=)     }
-  it { should respond_to(:stepExpireTime)    }
-  it { should respond_to(:stepExpireTime=)   }
-  it { should respond_to(:ownerShouldSign)   }
-  it { should respond_to(:ownerShouldSign=)  }
-  it { should respond_to(:orderedSignature)  }
-  it { should respond_to(:orderedSignature=) }
-  it { should respond_to(:emailSubject)      }
-  it { should respond_to(:emailSubject=)     }
-  it { should respond_to(:emailBody)         }
-  it { should respond_to(:emailBody=)        }
-  it { should respond_to(:documentsCount)    }
-  it { should respond_to(:documentsCount=)   }
-  it { should respond_to(:documentsPages)    }
-  it { should respond_to(:documentsPages=)   }
-  it { should respond_to(:recipients)        }
-  it { should respond_to(:recipients=)       }
+  it { should have_accessor(:id)               }
+  it { should have_accessor(:name)             }
+  it { should have_accessor(:ownerId)          }
+  it { should have_accessor(:ownerGuid)        }
+  it { should have_accessor(:reminderTime)     }
+  it { should have_accessor(:stepExpireTime)   }
+  it { should have_accessor(:ownerShouldSign)  }
+  it { should have_accessor(:orderedSignature) }
+  it { should have_accessor(:emailSubject)     }
+  it { should have_accessor(:emailBody)        }
+  it { should have_accessor(:documentsCount)   }
+  it { should have_accessor(:documentsPages)   }
+  it { should have_accessor(:recipients)       }
 
-  it { should have_alias(:owner_id, :ownerId)                     }
-  it { should have_alias(:owner_id=, :ownerId=)                   }
-  it { should have_alias(:owner_guid, :ownerGuid)                 }
-  it { should have_alias(:owner_guid=, :ownerGuid=)               }
-  it { should have_alias(:reminder_time, :reminderTime)           }
-  it { should have_alias(:reminder_time=, :reminderTime=)         }
-  it { should have_alias(:step_expire_time, :stepExpireTime)      }
-  it { should have_alias(:step_expire_time=, :stepExpireTime=)    }
+  it { should alias_accessor(:owner_id, :ownerId)                }
+  it { should alias_accessor(:owner_guid, :ownerGuid)            }
+  it { should alias_accessor(:reminder_time, :reminderTime)      }
+  it { should alias_accessor(:step_expire_time, :stepExpireTime) }
   # owner_should_sign is overwritten
-  it { should have_alias(:owner_should_sign=, :ownerShouldSign=)  }
+  it { should have_alias(:owner_should_sign=, :ownerShouldSign=) }
   # ordered_signature is overwritten
   it { should have_alias(:ordered_signature=, :orderedSignature=) }
-  it { should have_alias(:email_subject, :emailSubject)           }
-  it { should have_alias(:email_subject=, :emailSubject=)         }
-  it { should have_alias(:email_body, :emailBody)                 }
-  it { should have_alias(:email_body=, :emailBody=)               }
-  it { should have_alias(:documents_count, :documentsCount)       }
-  it { should have_alias(:documents_count=, :documentsCount=)     }
-  it { should have_alias(:documents_pages, :documentsPages)       }
-  it { should have_alias(:documents_pages=, :documentsPages=)     }
+  it { should alias_accessor(:email_subject, :emailSubject)       }
+  it { should alias_accessor(:email_body, :emailBody)             }
+  it { should alias_accessor(:documents_count, :documentsCount)   }
+  it { should alias_accessor(:documents_pages, :documentsPages)   }
 
   describe '#recipients=' do
     it 'converts each recipient to GroupDocs::Signature::Recipient object if hash is passed' do